KP-8598 Implemented ability to delete draft requests
Added a delete button to the form page when viewing a draft request.
Added a delete button to the requests lists that appears on hover or focus of a draft row.
Added swipe delete functionality for mobile devices that allows swiping left on a draft record in the requests list to delete it.
All of the above deletes show a confirmation popup.
Implemented a useSwipe hook to handle swiping logic.