KP-9019 Added settings page
Added a settings page with links to the profile, datastore, and any form of type Settings from the admin-center kapp.
Added a datastore UI that shows forms of type Data from the admin-center kapp and allows for seeing their submissions and rendering the forms.
Updated header to have a dropdown with links to the profile and settings page.